Started working on step module
