+++ backend/src/app.module.ts
@@ -2,10 +2,12 @@ import { Module } from '@nestjs/common';
 import { BackendController } from './app.controller';
 import { BackendService } from './app.service';
 import { FpbModule } from './routes/fpb/fpb.module';
+import { StepModule } from './routes/step/step.module';
 
 @Module({
 	imports: [
-		FpbModule
+		// FpbModule,
+		StepModule
 	],
 	controllers: [BackendController],
 	providers: [BackendService],
