+++ backend/src/app.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 @Injectable()
 export class BackendService {
-  getHello(): string {
-    return 'Hello World!';
-  }
+	getHello(): string {
+		return 'Hello World!';
+	}
 }
