+++ backend/src/main.ts
@@ -1,8 +1,8 @@
 import { NestFactory } from '@nestjs/core';
-import { BackendModule } from './backend.module';
+import { BackendModule } from './app.module';
 
 async function bootstrap() {
-  const app = await NestFactory.create(BackendModule);
-  await app.listen(3000);
+	const app = await NestFactory.create(BackendModule);
+	await app.listen(3000);
 }
 bootstrap();
