+++ backend/src/routes/step/step.controller.ts
@@ -1,7 +1,7 @@
 import { Controller } from '@nestjs/common';
 
 @Controller()
-export class BackendController {
+export class StepController {
 	constructor() {}
 
 }
