+++ backend/src/routes/step/step.module.ts
@@ -1,0 +1,14 @@
+import { Module } from '@nestjs/common';
+import { StepController } from './step.controller';
+import { StepService } from './step.service';
+
+@Module({
+	imports: [],
+	controllers: [
+		StepController
+	],
+	providers: [
+		StepService
+	],
+})
+export class StepModule {}
